Revert "Separated nested classes in to their own files & added namespace"
