--- Library/AttributeValues.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 882b29f3f77fbb7449d6c8b5d590bb51
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
