--- Library/AttributeValues/AttributeDefinition.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eb38c0e1197a7ff45b90079b15c7e7c9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
