--- Library/AttributeValues/AttributeType.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d2a46a846e91b924db9e529909c58d00
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
