--- Library/AttributeValues/AttributeValue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1eecb0f05ede83a41b9dd926259f760a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
