--- Library/AttributeValues/FloatAttributeValue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c170727a11659dd4e86330352ddb64bc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
