--- Library/AttributeValues/IntAttributeValue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e8aac72f38ea4b04eb7c97beefecdda8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
