--- Library/Edge.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 36b8b26013e30b147a2291290822098e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
