--- Library/Face.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9d134271c7423114abcff8f4ad3fa702
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
