--- Library/Loop.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e70c0bd3b33d17945bf0a6ac0a58763d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
