--- Library/Vertex.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d10d9513658db3548b4967d704928c89
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
