+++ Tests/BMeshUnityTests.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
-using BMeshLib.Tests;
 
 namespace Tests
 {
